% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.R
\name{decompress_fst}
\alias{decompress_fst}
\title{Decompress a raw vector with compressed data.}
\usage{
decompress_fst(x)
}
\arguments{
\item{x}{raw vector with data previously compressed with \code{compress_fst}.}
}
\value{
a raw vector with previously compressed data.
}
\description{
Decompress a raw vector with compressed data.
}
