% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.rbind.R
\name{fst.rbind}
\alias{fst.rbind}
\title{Add rows to the data frame stored in a \code{fst} file.}
\usage{
fst.rbind(path, x, compress = NULL)
}
\arguments{
\item{path}{Path to a \code{fst} file}

\item{x}{A data frame to append to an existing \code{fst} file.}

\item{compress}{Value in the range 0 to 100, indicating the amount of compression to use for the appended data frame.
If \code{NULL}, the compression setting of the existing \code{fst} file will be used.}
}
\description{
Take an existing \code{fst} file and append rows from a (in-memory) table.
}
