% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst_table.R
\name{fst}
\alias{fst}
\title{Access a fst file like a data frame}
\usage{
fst(path, old_format = FALSE)
}
\arguments{
\item{path}{path to fst file}

\item{old_format}{use TRUE to read fst files generated with a fst package version lower than v0.8.0}
}
\value{
An object of class \code{fst_table}
}
\description{
These functions permit using a dataset stored in a fst file much like a regular
data frame.
}
