% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstcompress.R
\name{fstdecompress}
\alias{fstdecompress}
\title{Decompress a raw vector with compressed data.}
\usage{
fstdecompress(x)
}
\arguments{
\item{x}{raw vector with data previously compressed with \code{fstcompress}.}
}
\value{
a raw vector with previously compressed data.
}
\description{
Decompress a raw vector with compressed data.
}
