% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsthash.R
\name{fsthash}
\alias{fsthash}
\title{Parallel calculation of the hash of a raw vector}
\usage{
fsthash(x, seed = NULL)
}
\arguments{
\item{x}{raw vector that you want to hash}

\item{seed}{The seed value for the hashing algorithm. If NULL, a default seed will be used.}
}
\value{
hash value
}
\description{
Parallel calculation of the hash of a raw vector
}
