% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{fstmeta}
\alias{fstmeta}
\alias{fst.metadata}
\title{Read metadata from a fst file}
\usage{
fstmeta(path)

fst.metadata(path)
}
\arguments{
\item{path}{path to fst file}
}
\value{
Returns a list with meta information on the stored dataset in \code{path}.
Has class 'fstmetadata'.
}
\description{
Method for checking basic properties of the dataset stored in \code{path}.
}
