% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmp.R
\name{getFstThreads}
\alias{getFstThreads}
\title{Get the number of threads use by all OpenMP parallel calculations in \code{fst}}
\usage{
getFstThreads()
}
\value{
The number of threads used by OpenMP.
}
\description{
Get the number of threads use by all OpenMP parallel calculations in \code{fst}
}
