% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstcompress.R
\name{fstcompress}
\alias{fstcompress}
\title{Compress a raw vector using the LZ4 or ZSTD compressor.}
\usage{
fstcompress(x, compressor = "LZ4", compression = 0)
}
\arguments{
\item{x}{raw vector.}

\item{compressor}{compressor to use for compressing \code{x}. Valid options are "LZ4" (default) and "ZSTD".}

\item{compression}{compression factor used. Must be in the range 0 (lowest compression) to 100 (maximum compression).}
}
\description{
Compress a raw vector using the LZ4 or ZSTD compressor.
}
