% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmp.R
\name{setFstThreads}
\alias{setFstThreads}
\title{Set the number of threads to use in OpenMP parallel calculations in \code{fst}}
\usage{
setFstThreads(nrOfThreads)
}
\value{
The previous number of threads set.
}
\description{
Set the number of threads to use in OpenMP parallel calculations in \code{fst}
}
